# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GCC-TBC Pro Typing Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /certificates:
    get:
      operationId: listCertificates
      tags: [certificates]
      summary: List certificates
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: language
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Certificates list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateListResponse"

    post:
      operationId: generateCertificate
      tags: [certificates]
      summary: Generate a certificate for a passing result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateInput"
      responses:
        "201":
          description: Generated certificate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
  /certificates/{id}:
    get:
      operationId: getCertificate
      tags: [certificates]
      summary: Get certificate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Certificate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
  /certificates/verify/{verificationId}:
    get:
      operationId: verifyCertificate
      tags: [certificates]
      summary: Verify a certificate by verification ID (public)
      parameters:
        - name: verificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Certificate verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateVerification"
        "404":
          description: Invalid certificate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── SUBSCRIPTIONS ────────────────────────────────────────────────────────────
...
  /commissions/my:
    get:
      operationId: listMyCommissions
      tags: [commissions]
      summary: List commissions for the current institute admin's own institute
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Commissions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionListResponse"
  /commissions:
    get:
      operationId: listCommissions
      tags: [commissions]
      summary: List all commissions across all institutes (owner only)
      parameters:
        - name: instituteId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: fromDate
          in: query
          schema:
            type: string
            format: date-time
        - name: toDate
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Commissions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionListResponse"
  /commissions/{id}:
    patch:
      operationId: updateCommissionStatus
      tags: [commissions]
      summary: Transition a commission's status (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommissionStatusUpdate"
      responses:
        "200":
          description: Updated commission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commission"

  # ─── PAYMENTS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Certificate:
      type: object
      required: [id, userId, resultId, certificateNumber, verificationId, language, netWpm, accuracy, issuedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        resultId:
          type: integer
        certificateNumber:
          type: string
        verificationId:
          type: string
        userName:
          type: string
        language:
          type: string
        speedCategory:
          type: integer
        netWpm:
          type: number
        accuracy:
          type: number
        instituteName:
          type: ["string", "null"]
        issuedAt:
          type: string
          format: date-time
    CertificateInput:
      type: object
      required: [resultId]
      properties:
        resultId:
          type: integer
    CertificateListResponse:
      type: object
      required: [certificates, total, page, limit]
      properties:
        certificates:
          type: array
          items:
            $ref: "#/components/schemas/Certificate"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CertificateVerification:
      type: object
      required: [valid, certificate]
      properties:
        valid:
          type: boolean
        certificate:
          $ref: "#/components/schemas/Certificate"

    # ─── SUBSCRIPTION SCHEMAS ─────────────────────────────────────────────────
...
    Commission:
      type: object
      required: [id, instituteId, paymentId, studentUserId, amountInPaise, commissionType, commissionRate, status, createdAt]
      properties:
        id:
          type: integer
        instituteId:
          type: integer
        paymentId:
          type: integer
        studentUserId:
          type: integer
        amountInPaise:
          type: integer
        commissionType:
          type: string
          enum: [percent, flat]
        commissionRate:
          type: number
        status:
          type: string
          enum: [pending, approved, paid]
        paidAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    CommissionStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, approved, paid]
    CommissionListResponse:
      type: object
      required: [commissions, total, page, limit]
      properties:
        commissions:
          type: array
          items:
            $ref: "#/components/schemas/Commission"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── HIERARCHY-AWARE REGISTRATION SCHEMAS ────────────────────────────────
...